+++ server/src/bin.rs
@@ -12,6 +12,7 @@ mod helpers;
 #[cfg(feature = "https")]
 mod https;
 mod jsonerrors;
+#[cfg(feature = "process-management")]
 mod process;
 mod routes;
 pub mod serve;
@@ -58,6 +59,25 @@ async fn main() -> errors::AtomicServerResult<()> {
             appstate.store.import(&readstring)?;
 
             println!("Sucesfully imported {:?} to store.", o.path);
+            Ok(())
+        }
+        Some(config::Command::ShowConfig) => {
+            println!("{:#?}", config);
+            Ok(())
+        }
+        Some(config::Command::Reset) => {
+            if dialoguer::Confirm::with_theme(&dialoguer::theme::ColorfulTheme::default())
+                .with_prompt(
+                    format!("Warning!! Do you really want to remove all data from your atomic-server? This will delete {:?}", &config.store_path),
+                )
+                .interact()
+                .unwrap()
+            {
+                std::fs::remove_dir_all(config.store_path).map(|e| format!("unable to remove directory: {:?}", e))?;
+                println!("Done");
+            } else {
+                println!("Ok, not removing anything.");
+            }
             Ok(())
         }
         Some(config::Command::SetupEnv) => {