+++ CHANGELOG.md
@@ -5,9 +5,16 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 
 ## UNRELEASED
 
+- **Warning**: Various default directories have moved (see #331). Most notably the `data` directory. The location depends on your OS. Run `show-config` to see where it will be stored now. If you have data in `~/.config/atomic/db`, move it to this new directory.
 - **Warning**: Search index will have to be rebuilt. Start with `--rebuild-index`.
 - Updated various dependencies, and made `cargo.toml` less restrictive.
 - Remove `async-std` calls from `upload.rs`
+- Added `reset` and `show-config` commands to `atomic-server`.
+- Added `data-dir` flag
+- Replaced `awc` with `ureq` #374
+- Get rid of `.unwrap` calls in `commit_monitor` #345
+- Make process management optional #324 #334
+- Auto-update desktop distributions using Tauri #158
 
 ## [v0.31.1] - 2022-03-29